Refactor parse table decoder
